+++ Example/SideMenuControllerExample/Home/Services/DataLoader.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-struct DataLoader {
+struct DataLoader: Sendable {
 
     enum DataLoaderError: Error {
         case fileDoesNotExist
@@ -33,5 +33,4 @@ struct DataLoader {
 
         return decodedData
     }
-
 }