+++ Sources/JESideMenuController/Protocols/LayoutBuildingProtocol.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-protocol LayoutBuilding {
+protocol LayoutBuilding: Sendable {
 
     /// Visible spacing of the content view if the menu is open
     var spacing: CGFloat { get }
@@ -18,8 +18,7 @@ protocol LayoutBuilding {
     /// Builds the layout for a slider menu in the specified view.
     /// - parameter view: The entry point to build up the view hierarchy.
     /// - parameter isLeft: A Boolean value that determines the side which the menu will be placed.
-    func layout(in view: UIView?, isLeft: Bool)
-
+    @MainActor func layout(in view: UIView?, isLeft: Bool)
 }
 
 extension LayoutBuilding {
@@ -36,23 +35,28 @@ extension LayoutBuilding {
             scrollView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: spacing).isActive = true
         }
     }
-
-    /// Calculates the width of the scroll view based on a given size and the current devide (iPhone/iPad)
-    func getScrollViewWidth(for size: CGSize,
-                            userInterfaceIdiom: UIUserInterfaceIdiom = UIDevice.current.userInterfaceIdiom) -> CGFloat {
-        if userInterfaceIdiom == .pad {
-            return ipadWidth
-        } else {
-            return size.width - spacing
-        }
-    }
-
 }
 
-// Enable testing for `getScrollViewWidth(for:, userInterfaceIdiom:)`
-struct DefaultLayoutBuilder: LayoutBuilding {
-    var spacing: CGFloat
-    var ipadWidth: CGFloat
+/// Utility to calculate the scrollable for a given layout.
+struct LayoutUtil: LayoutBuilding {
+
+    /// The visible spacing for the content view when the menu is open
+    let spacing: CGFloat
+
+    /// The width of the menu on iPad
+    let ipadWidth: CGFloat
 
     func layout(in view: UIView?, isLeft: Bool) {}
+
+    /// Calculates the width of the scroll view based on a given size and the current device (iPhone/iPad)
+    func getScrollViewWidth(
+        for size: CGSize,
+        userInterfaceIdiom: UIUserInterfaceIdiom = UIDevice.current.userInterfaceIdiom
+    ) -> CGFloat {
+        if userInterfaceIdiom == .pad {
+            ipadWidth
+        } else {
+            size.width - spacing
+        }
+    }
 }