+++ Tests/JESideMenuControllerTests/UIViewControllerExtensionTests.swift
@@ -9,15 +9,8 @@
 import XCTest
 @testable import JESideMenuController
 
-class UIViewControllerExtensionTests: XCTestCase {
-
-    override func setUp() {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
+@MainActor
+final class UIViewControllerExtensionTests: XCTestCase {
 
     func testSideMenuControllerExtension() {
         // Given
@@ -52,5 +45,4 @@ class UIViewControllerExtensionTests: XCTestCase {
         parent.remove(controller: viewcontroller)
         XCTAssertEqual(parent.children.count, 0)
     }
-
 }