+++ CHANGELOG.md
@@ -7,16 +7,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 ### Added
 - Documenting changes in CHANGELOG file.
 - Increased code coverage.
-- Added functionality to adjust the side menu configuration. (spacing, iPad menu width, drop shadow image etc.)
+- Feature to enable/disable scolling
+- Added functionality to adjust the side menu configuration. (spacing, iPad menu width, drop shadow image, tintColor etc.)
+- Added Dark Mode support
 
 ### Changed
 - Improved example application.
 - Switched from Travis-CI to github actions.
 - Require minimum target iOS 15
+- Updated example project
+- Added @MainActor annotation and Sendable conformance
 
 ### Fixed
 - Rewording of description and fix typos in README.
-- Fixed a bug where the slide-in edge gesture recognizer would block the recognition of UINavigationControllers back swipe.
+- Fixed a bug where the slide-in edge gesture recognizer would block the recognition of `UINavigationControllers` back swipe.
 
 ## [1.0.0] - 2019-07-26
 ### Added